// src/app/core/models/masters/incident-management/incident-damage-type.ts
export interface IncidentDamageTypes {
    id: number;
    title: string;
    description: string;
    status: string;
    status_id :number;
    status_label: string;
}
export interface IncidentDamageTypesPaginationResponse {
    current_page: number;
    total: number;
    per_page: number;
    last_page: number;
    from: number;
    data: IncidentDamageTypes[];
}